+++ src/App.css
@@ -39,6 +39,7 @@
 }
 .sizeHover:hover{
     font-size: 18px;
+    cursor: pointer;
 }
 .rowAppenant{
     width: 100%;
@@ -46,10 +47,10 @@
 }
 .colApprenant{
     display: flex;
-    flex-direction: column;
+    /* flex-direction: column; */
     /* flex-wrap: wrap; */
     /* justify-content: center; */
-    align-items: center;
+    /* align-items: center; */
 }
 .cardSpan{
     flex-wrap: wrap;
@@ -58,8 +59,8 @@
     display: flex;
     flex-wrap: wrap;
     flex-direction: column;
-    justify-content: center;
-    align-items: center;
+    /* justify-content: center; */
+    /* align-items: center; */
     justify-content: space-around;
 }
 .icon {
